Fixes potential leaks caused by disjoint 'using' blocks.
Implements a UseSafely extension method to guarantee correct cleanup of
disposables when the source disposable throws during Dispose().

LargeCollectionSorter passed the new test before this change, but this is
a limitation of the test and it's not obvious why. The source enumerator
is entirely consumed during batching and the batch resources are guarded
specifically, so a failing test would need to trigger random failures in
the Dispose() methods of the accumulators constructed inside the Sorter.
A failure when disposing the source is not sufficient to trigger the bug
in this case.